Reticule is partially done
We now have a reticule on the mouse cursor. Note it will throw an
exception due to the player not having a weapon for it to get a type
from. I'll be working on getting our first actual reticule done tonight,
though it won't work until we get the weapon manager up and running