Room generation stable
Room generation will now functionally generate 10-20+ rooms.

In rare cases, a door can go missing and a path to the outside is
opened. I think this has to do with adjacent rooms interacting poorly.
I'll try to fix that later, but it is sufficiently rare I think it can
be safely ignored.